Add Nowcast lightning (#611)
* Tweaks to make lightning nowcast work better. These may be superceded by other PRs.

* First step in adding lightning nowcast with tests (not enough unit tests yet)

* Added unit tests for everything and fixed bugs where stuff didn't work.

* Adds docstring and unit tests for rescale function.

* Improvements to Pylint score.

* Improvements for Pylint.

* Improves PEP8

* Tweaks to make lightning nowcast work better. These may be superceded by other PRs.

* First step in adding lightning nowcast with tests (not enough unit tests yet)

* Added unit tests for everything and fixed bugs where stuff didn't work.

* Improves PEP8

* Shortened lines to 80-char for PEP8

* Modifies arguments to rescale function.

* Fixes a couple of issues following a branch-rebase.
Replaces rescale function with improver.utilities.rescale.

* Reverts edits to non-nowcast modules as these broke unit tests.

* Modifies unittest cubes to have useful forecast-period coords.

* Adds a range check to the rescaling function and unit tests for this.

* Tweaks to make branch PEP8 and Pylint compliant following rebase.

* Adjusts process_haloes so that is can use the NeighbourhoodProcessing ability to vary the radius with lead time.

* Tweaks threshold as it is now stored as a float, not an int.

* Improves doc-strings

* Moves some hard-coded values to class.__init__

* Fixed bug in unit-test

* All hard-coded values are now optional configuration items for the plugin with default values.

* Updates following Gavin's review in an internal Met Office ticket.

* Restructures docstrings to make them Sphinx-compliant.

* Changes following Gavin's second review.

* Refactors code for limiting prob(lightning) with rescaled prob(precip) so it can be reused.

* Adds code to plugin to make use of VII data and modifies existing unit tests accordingly.

* Adds simple tests using VII data.

* Changes VII cube to expect probabilities. Adds unit tests.

* Adds unit test for process with VII.

* Adds a useful comment.

* One last change for using rescale from rebased branch.

* Adds unit tests to make sure that a VII cube is correctly picked up and used.

* Refactors plugin test code for setting up VII tests.

* Renames new setup function

* Adds missing documentation for new VII parameters to the plugin.

* Adds code to increase prob(lightning) where precipitation rates are very high.

* Updates existing unit-tests to work with precipitation data with three thresholds.

* Removes threshold coordinate from output cube.

* Adds unit-tests for heavy and intense precipitation.

* Adds null tests for heavy and intense precipitation

* Changes plugin test so that VII data increases along y-axis

* Adds precip-rate increasing along x-axis giving 4x4 results to test.

* PEP8

* PYLINT improvements

* Refactors handling of radar ice (VII) data to separate plugin.

* Refactor code so ApplyPrecip is a separate plugin and apply_double_scaling is with rescale.

* PEP8

* PYLINT

* Result of running 2to3 on nowcast code.

* Updated code for Iris 2.0 and updated licences. All tests pass now.

* Moves lightning and convection plugins from nowcast to nowcasting.

* Updated initialisation docstrings and added a few comments for clarity.

* Moves only function from handle_precip.py into NowcastLightning plugin along with associated unit tests.

* Moves only function from handle_vii.py into NowcastLightning plugin along with associated unit tests.

* First set of post-review changes.

* Second set of post-review changes. Standardises definition of LR2 which modifies one of the unit-tests.

* Third set of post-review changes to increase the clarity of the code.

* Fix errors caused by Plugin changes.

* Standardises test cube definitions and adds documentation describing them.

* Improves docstring for unit test for apply_double_scaling.

* Improves doc-strings and comments for new utilities function and unit-tests.

* Updates unit-tests in final response to Caroline's first review.

* Improves checks for missing inputs and adds unit tests for these.

* Responds to more review comments.

* Improves doc-strings a little more.

* For Pycodestyle

* For Codacy and Travis (Uses raw-strings for defining pattern-matches for Regexp that include backslash).

* Responds to more review comments.

* Constraints modified to test for nearly-equal floating-point values. Removed some unused values picked up by pylint.

* Improves clarity further following Caroline's latest review.

* Moves initiation of plugin to setUp functions.

* Ensures coord data type is float32, not float64.

* Removed unnecessary unit-test

* Adjusts indentation in doc-strings.

* Further improvements to comments.

* Improves doc-string and variable name following Fiona's review.

* Pycodestyle change.

* Updates doc-strings for content following Fiona's review.
Updates doc-strings for Sphinx formatting.
Adds insistence that precipitation cube threshold coordinate has the correct units and updates the unit test that doesn't have this coord as it now fails earlier in the process.

* Further improvements for Sphinx

* Lots of little tweaks following some of Fiona's review. More to follow.

* Improved repr method to print class name which allowed unit-test to be simplified.

* Simplified all unit-tests except Test_process to use 3 x 3 horizontal data arrays instead of 16 x 16. Makes debugging and comprehension easier.

* Reduces horizontal coordinate lengths from 16 to 4 to make these unit tests easier to debug and understand.

* Unit-test improvements.

* Removes warnings from Sphinx about unexpected indentations.

* Makes ice_thresholds visible in Sphinx documentation so that they can be seen by users.

* Removes duplicate import. Improves Sphinx-rendering of doc-string.

+++ lib/improver/tests/nbhood/nbhood/test_BaseNeighbourhoodProcessing.py
@@ -133,7 +133,8 @@ def set_up_cube(zero_point_indices=((0, 0, 7, 7),), num_time_points=1,
                                 standard_name="time", units=tunit), 1)
 
     step_size = 2000
-    y_points = np.arange(0, step_size*num_grid_points, step_size)
+    y_points = np.arange(0., step_size*num_grid_points, step_size,
+                         dtype=np.float32)
     cube.add_dim_coord(
         DimCoord(
             y_points,
@@ -144,7 +145,8 @@ def set_up_cube(zero_point_indices=((0, 0, 7, 7),), num_time_points=1,
         2
     )
 
-    x_points = np.arange(-50000, (step_size*num_grid_points)-50000, step_size)
+    x_points = np.arange(-50000., (step_size*num_grid_points)-50000, step_size,
+                         dtype=np.float32)
     cube.add_dim_coord(
         DimCoord(
             x_points,
